Remove Req-SingleRedemption
We have discussed a similar requirement during the interim IETF meeting. It was discussed that the transfer protocol should not be in charge of ensuring the provisioning information is only redeemed once. This responsibility is up to the credential authority. Suggesting an edit to change it to ensure the Sender and Receiver can communicate privately without a third party intercepting the messages.